change randomsingleton to have seed as unsigned long

+++ python/cRandomSingleton.pxd
@@ -1,10 +1,4 @@
 cdef extern from "util/RandomSingleton.h":
-    cdef void randomsingleton_seed "RandomSingleton::seed"(unsigned int seed)
+    cdef void randomsingleton_seed "RandomSingleton::seed"(unsigned long seed)
     cdef float randomsingleton_uniform "RandomSingleton::uniform"()
-
-#    cdef cppclass RandomSingleton:
-#        @staticmethod
-#        void seed(unsigned int seed)
-#        @staticmethod
-#        float uniform()
